Simple solution of most frequently asked question
Here is the most simplest possible solution of this problem. 
The wording of the problem is confuses. But logically if we think the sum of any given integer list would be minimum if start decreasing the maximum numbers. 

The example tries to confuse as well. We we read it carefully, we understand the way to solve it .